fix: ios subscription type casting (#519)
Fixed an issue where the **`subscriptionPeriodNumberIOS`** field was
being returned as an integer when fetching subscription product
information on iOS, which caused type casting errors in Flutter when
converting it to a string. Updated the Swift code to convert this value
to a string. Additionally, simplified the names of unused stream
controllers in Flutter to **`_purchaseUpdatedListener`** and
**`_purchaseErrorListener`**, and removed the unnecessary
**`Flutter.podspec`** file.